+++ Vm/vm.h
@@ -6,7 +6,7 @@
 /*   By: ahamouda <ahamouda@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/06/04 22:30:40 by ahamouda          #+#    #+#             */
-/*   Updated: 2016/08/30 17:10:41 by ahamouda         ###   ########.fr       */
+/*   Updated: 2016/08/30 22:48:39 by ahamouda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -70,12 +70,12 @@ void		introduce_players(t_vm_data *arena);
 ** ACCESSOR FUNCTIONS
 */
 
-int			get_n_param_value(t_vm_data *arena, t_proc *proc, int nb, int *err);
+long		get_n_param_value(t_vm_data *arena, t_proc *proc, int nb, int *err);
 int			get_n_param_size(t_vm_data *arena, int n);
 int			get_n_reg_param_value(t_vm_data *arena, t_proc *process, int nb,
 			int *error);
 void		get_number_of_players(t_vm_data *arena);
-int			get_param_value(t_vm_data *arena, int pc, int size);
+long		get_param_value(t_vm_data *arena, int pc, int size);
 int			get_param_type(t_vm_data *arena, int param_number);
 int			get_parameter_size(int op_code, int var_code);
 t_player	*get_winner(t_vm_data *arena);
